requirements.txt
#Django 
Django==4.1
django-configurations==2.4

#Django Rest Framework
djangorestframework==3.13.1
Markdown==3.3.7
django-filter==22.1

#Authentication
django-rest-authemail==2.1.4

#Documentation
mkdocs==1.3.1

#Tests
factory-boy==3.2.1
coverage== 6.4.4

#Database Connection
psycopg2==2.9.3

#image
pillow==9.2.0

#production server
django-storages==1.13.1
boto3==1.24.61
gunicorn==20.1.0


pydantic==1.10.2


